Split core.py into more files
