+++ python/lammps/__init__.py
@@ -1,2 +1,4 @@
+from .constants import *
 from .core import *
+from .data import *
 from .pylammps import *
